drop immedateAction, use existing way to drive user actions
