+++ src/game/building/farmersInn.js
@@ -1,0 +1,4 @@
+import { compose } from 'redux'
+import { actionOption } from '../common/player'
+
+export default state => compose(actionOption('farmersInn'))(state)
